# Backend/GradeProjectStage3/generated/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Api Documentation",
    "version" : "1.0",
    "title" : "Api Documentation",
    "termsOfService" : "urn:tos",
    "contact" : { },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host" : "localhost:8080",
  "basePath" : "/",
...
  "paths" : {
    "/admin/add-customer" : {
      "post" : {
        "tags" : [ "admin-controller" ],
        "summary" : "Add Customer",
        "description" : "Adds a Customer and returns it\n\n[id - unnecessary]",
        "operationId" : "addCustomerUsingPOST",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "customerReq",
          "description" : "customerReq",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Admin"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Succesful operation",
            "schema" : {
              "$ref" : "#/definitions/Customer"
            }
          },
          "201" : {
            "description" : "Created"
          },
          "400" : {
            "description" : "Customer addition failed",
            "schema" : {
              "$ref" : "#/definitions/ExceptionResponse"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
...
    "/admin/create-company" : {
      "post" : {
        "tags" : [ "admin-controller" ],
        "summary" : "Create Company",
        "description" : "Creates a Company and returns it\n\n[id - unnecessary]",
        "operationId" : "createCompanyUsingPOST",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "companyReq",
          "description" : "companyReq",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Admin"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Succesful operation",
            "schema" : {
              "$ref" : "#/definitions/Company"
            }
          },
          "201" : {
            "description" : "Created"
          },
          "400" : {
            "description" : "Company creation failed",
            "schema" : {
              "$ref" : "#/definitions/ExceptionResponse"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
...
    "/login-manager/login" : {
      "post" : {
        "tags" : [ "login-manager-controller" ],
        "summary" : "Login",
        "description" : "Login based on user type",
        "operationId" : "loginUsingPOST",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "email",
          "in" : "query",
          "description" : "Email of the user",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "password",
          "in" : "query",
          "description" : "Password of the user",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "userType",
          "in" : "query",
          "description" : "Type of the user",
          "required" : true,
          "type" : "string",
          "enum" : [ "ADMIN", "COMPANY", "CUSTOMER" ]
        } ],
        "responses" : {
          "200" : {
            "description" : "Succesful operation",
            "schema" : {
              "$ref" : "#/definitions/JwtResponse"
            }
          },
          "201" : {
            "description" : "Created"
          },
          "400" : {
            "description" : "Login failed",
            "schema" : {
              "$ref" : "#/definitions/ExceptionResponse"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/login-manager/register-company" : {
      "post" : {
        "tags" : [ "login-manager-controller" ],
        "summary" : "Register Company",
        "description" : "Registers a Company and returns it\n\n[id - unnecessary]",
        "operationId" : "createCompanyUsingPOST_1",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "companyReq",
          "description" : "companyReq",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Admin"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Succesful operation",
            "schema" : {
              "$ref" : "#/definitions/Company"
            }
          },
          "201" : {
            "description" : "Created"
          },
          "400" : {
            "description" : "Company creation failed",
            "schema" : {
              "$ref" : "#/definitions/ExceptionResponse"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/login-manager/register-customer" : {
      "post" : {
        "tags" : [ "login-manager-controller" ],
        "summary" : "Register Customer",
        "description" : "Registers a Customer and returns it\n\n[id - unnecessary]",
        "operationId" : "addCustomerUsingPOST_1",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "customerReq",
          "description" : "customerReq",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Admin"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Succesful operation",
            "schema" : {
              "$ref" : "#/definitions/Customer"
            }
          },
          "201" : {
            "description" : "Created"
          },
          "400" : {
            "description" : "Customer addition failed",
            "schema" : {
              "$ref" : "#/definitions/ExceptionResponse"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    }
  },
  "definitions" : {
    "Admin" : {
      "type" : "object",
      "properties" : {
        "email" : {
          "type" : "string",
          "example" : "email@email.com",
          "description" : "Min=1, Max=255 (spaces gets removed).\nMust be in email format.\nGets lower-cased."
        },
        "firstName" : {
          "type" : "string",
          "description" : "Min=1, Max=255 (gets trimmed)."
        },
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 0,
          "description" : "Auto-incremented."
        },
        "lastName" : {
          "type" : "string",
          "description" : "Min=1, Max=255 (gets trimmed)."
        },
        "password" : {
          "type" : "string",
          "description" : "Min=1, Max=255."
        }
      },
      "title" : "Admin",
      "description" : "Email between users must be unique.\n\nManages Companies.\nManages Customers\nManages Logs."
    },
    "Company" : {
      "type" : "object",
      "properties" : {
        "coupons" : {
          "type" : "array",
          "description" : "Only as response.",
          "items" : {
            "$ref" : "#/definitions/Coupon"
          }
        },
        "email" : {
          "type" : "string",
          "example" : "email@email.com",
          "description" : "Min=1, Max=255 (spaces gets removed).\nMust be in email format.\nGets lower-cased."
        },
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "Auto-incremented."
        },
        "name" : {
          "type" : "string",
          "description" : "Min=1, Max=255 (gets trimmed)."
        },
        "password" : {
          "type" : "string",
          "description" : "Min=1, Max=255."
        },
        "role" : {
          "type" : "string",
          "enum" : [ "ADMIN", "COMPANY", "CUSTOMER" ]
        }
      },
      "title" : "Company",
      "description" : "Email between users must be uinque.\nname between Companies must be unique.\n\nOwns and manages coupons."
    },
    "Coupon" : {
      "type" : "object",
      "properties" : {
        "amount" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "Min=0, Max=2147483647."
        },
        "category" : {
          "type" : "string",
          "enum" : [ "ATTRACTION", "AUTOMOBILE", "ELECTRICITY", "FOOD", "GAMING", "RESTAURANT", "SPORT", "VACATION" ]
        },
        "description" : {
          "type" : "string",
          "description" : "Min=1, Max=255 (gets trimmed)."
        },
        "endDate" : {
          "type" : "string",
          "format" : "date",
          "example" : "yyyy-mm-dd",
          "description" : "Cannot be in the past."
        },
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "Auto-incremented."
        },
        "image" : {
          "type" : "string",
          "description" : "Min=1, Max=255 (spaces gets removed)."
        },
        "price" : {
          "type" : "number",
          "format" : "double",
          "description" : "Min=0, Max=∞."
        },
        "startDate" : {
          "type" : "string",
          "format" : "date",
          "example" : "yyyy-mm-dd",
          "description" : "Cannot be after \"endDate\"."
        },
        "title" : {
          "type" : "string",
          "description" : "Min=1, Max=255 (gets trimmed)."
        }
      },
      "title" : "Coupon",
      "description" : "Title within the same Company must be unique.\n\nOwned and managed by Company.\nOwned by Customer."
    },
...
    "Customer" : {
      "type" : "object",
      "properties" : {
        "coupons" : {
          "type" : "array",
          "description" : "Only as response.",
          "items" : {
            "$ref" : "#/definitions/Coupon"
          }
        },
        "email" : {
          "type" : "string",
          "example" : "email@email.com",
          "description" : "Min=1, Max=255 (spaces gets removed).\nMust be in email format.\nGets lower-cased."
        },
        "firstName" : {
          "type" : "string",
          "description" : "Min=1, Max=255 (gets trimmed)."
        },
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "Auto-incremented."
        },
        "lastName" : {
          "type" : "string",
          "description" : "Min=1, Max=255 (gets trimmed)."
        },
        "password" : {
          "type" : "string",
          "description" : "Min=1, Max=255."
        },
        "role" : {
          "type" : "string",
          "enum" : [ "ADMIN", "COMPANY", "CUSTOMER" ]
        }
      },
      "title" : "Customer",
      "description" : "Email between users must be uinque.\n\nOwns and purchases coupons."
    },
    "ExceptionResponse" : {
      "type" : "object",
      "properties" : {
        "errorCode" : {
          "type" : "string",
          "example" : "CMP-001.001",
          "description" : "Exception error code."
        },
        "response" : {
          "type" : "string",
          "description" : "Exception message."
        }
      },
      "title" : "ExceptionResponse",
      "description" : "Throws exceptions, for unsuccessful responses.\n\nException List:\r\n\nexception:\r\nCustomExceptionList - LST-000.000\r\n\r\ncompanyException:\r\nCompanyAlreadyExistsException - CMP-001.001\r\nCompanyDuplicateValueException - CMP-001.002\r\nCompanyIsEmptyException - CMP-001.003\r\nCompanyIsNullException - CMP-001.004\r\nCompanyNotFoundException - CMP-001.005\r\nCompanyNullValueException - CMP-001.006\r\n\r\ncouponException:\r\nCouponAlreadyExistsException - CPN-002.001\r\nCouponDuplicateValueException - CPN-002.002\r\nCouponExpiredModifyException - CPN-002.003\r\nCouponExpiredPurchaseException - CPN-002.004\r\nCouponIsEmptyException - CPN-002.005\r\nCouponIsNullException - CPN-002.006\r\nCouponModifyNoStockException - CPN-002.007\r\nCouponNegativePriceException - CPN-002.008\r\nCouponNotFoundException - CPN-002.009\r\nCouponNullValueException - CPN-002.010\r\nCouponOwnedException - CPN-002.011\r\nCouponPurchaseNoStockException - CPN-002.012\r\n\r\ncustomerException:\r\nCustomerAlreadyExistsException - CST-003.001\r\nCustomerDuplicateValueException- CST-003.002\r\nCustomerIsEmptyException - CST-003.003\r\nCustomerIsNullException - CST-003.004\r\nCustomerNotFoundException - CST-003.005\r\nCustomerNullValueException - CST-003.006\r\n\r\nloggerException:\r\nLoggerIsEmptyException - LGR-004.001\r\n\r\nloginException:\r\nInvalidLoginException - LGN-005.001\r\nNullLoginException - LGN-005.002\r\nMismatchLoginException - LGN-005.003\r\n\r\nUserEmailAlreadyExists - USR-006.001"
    },
    "GrantedAuthority" : {
      "type" : "object",
      "properties" : {
        "authority" : {
          "type" : "string"
        }
      },
      "title" : "GrantedAuthority"
    },
    "JwtResponse" : {
      "type" : "object",
      "properties" : {
        "authorities" : {
          "type" : "array",
          "description" : "Authorities of the user.\n\nAuthority List:\r\n\nCUSTOMER\r\nCOMPANY\r\nADMIN",
          "items" : {
            "$ref" : "#/definitions/GrantedAuthority"
          }
        },
        "token" : {
          "type" : "string",
          "description" : "Jwt token of the user."
        },
        "type" : {
          "type" : "string",
          "example" : "Bearer",
          "description" : "Will always be \"Bearer\"."
        },
        "username" : {
          "type" : "string",
          "description" : "Email of the user."
        }
      },
      "title" : "JwtResponse",
      "description" : "Jwt response for the Jwt token."
    },
...
  }
}